requirements.txt
flask
langchain
langchain-openai
faiss-cpu
python-dotenv
pypdf
tiktoken
